Prevent file parsing from making network requests
When we were invoking the rdflib.Graph.parse() method, we were passing a
file-like object to the `source` parameter by position, which is documented as
accepting file-like objects as one of the possible formats.

However, using the `file` named parameter instead of `source` results in only
the file being parsed, and doesn't generate a network request.

The documentation in rdflib is far from clear on the difference between the
two parameters, but it seems to work, so let's roll with it.

Fixes #2.

Signed-off-by: Dan Scott <dan@coffeecode.net>