On second thought, let's not confuse sameAs with prior URIs
We saw a lot of the previous URIs just point to generic web pages for a system,
vs. a lot of the current URIs are specific for a given branch. A part is not
really the same as the whole. (But that would make for an opportunity for
parentOrganization properties, hmm...)

Signed-off-by: Dan Scott <dan@coffeecode.net>